Make Value attribute optional. Add more acceptance tests
